+++ src/main/java/com/telepathicgrunt/the_bumblezone/modinit/BzSounds.java
@@ -1,0 +1,16 @@
+package com.telepathicgrunt.the_bumblezone.modinit;
+
+import com.telepathicgrunt.the_bumblezone.Bumblezone;
+import com.telepathicgrunt.the_bumblezone.enchantments.CombCutterEnchantment;
+import net.minecraft.enchantment.Enchantment;
+import net.minecraft.util.ResourceLocation;
+import net.minecraft.util.SoundEvent;
+import net.minecraftforge.fml.RegistryObject;
+import net.minecraftforge.registries.DeferredRegister;
+import net.minecraftforge.registries.ForgeRegistries;
+
+public class BzSounds {
+    public static final DeferredRegister<SoundEvent> SOUND_EVENTS = DeferredRegister.create(ForgeRegistries.SOUND_EVENTS, Bumblezone.MODID);
+
+    public static final RegistryObject<SoundEvent> ANGERED_BEES = SOUND_EVENTS.register("angered_bees", () -> new SoundEvent(new ResourceLocation(Bumblezone.MODID, "angered_bees")));
+}